+++ src/index.ts
@@ -1,4 +1,4 @@
-export { bindLocale as applyLocale, isChainBuilder } from "./bindLocale";
+export { bindLocale, isChainBuilder } from "./bindLocale";
 export { ChainBuilder, createI18n } from "./chainBuilder";
 export { I18nMessage, isI18nMessage } from "./message";
 export type { LocalizedMessage } from "./message";